backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.25.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
weasyprint>=60.1
markdown>=3.5.0
python-dotenv>=1.0.0
reportlab>=4.0.0
